AkiraChix/requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
